Rename `exported_futures` modules to `storage`
Update the name to reflect that it doesn't contain any exported futures
anymore.
