+++ linera-sdk/src/lib.rs
@@ -50,10 +50,7 @@ pub mod test;
 pub mod util;
 pub mod views;
 
-use crate::{
-    contract::exported_futures::ContractStateStorage,
-    service::exported_futures::ServiceStateStorage,
-};
+use self::contract::ContractStateStorage;
 use async_trait::async_trait;
 use linera_base::{
     abi::{ContractAbi, ServiceAbi, WithContractAbi, WithServiceAbi},
@@ -66,6 +63,7 @@ use std::{error::Error, fmt::Debug, sync::Arc};
 pub use self::{
     extensions::{FromBcsBytes, ToBcsBytes},
     log::{ContractLogger, ServiceLogger},
+    service::ServiceStateStorage,
 };
 pub use linera_base::{abi, ensure};
 #[doc(hidden)]
